Add files via upload
A Caesar cipher is a simple method of encrypting messages used by Gaius Julius Caesar. It works by replacing a plaintext with a letter, some *fixed* number of positions down the alphabet.

For example, with a shift of 3, A would be replaced by D, B would become E, and so on.